Move status args to device verb
